Fixed motor speed bug
First attribute was being interpreted as a character and not a uint8_t
as it should. Turns out, I needed to use to_string to make it work

+++ Sockets/client/clientMain.cpp
@@ -208,7 +208,7 @@ void clientFrame::OnSetSpeedButtonClick(wxCommandEvent&)
     delete request;
     request = new struct MotorRequest;
     request->command = SPEED_CMD;
-    request->attributes[0] = speed;
+    request->attributes[0] = (uint8_t)speed;
     request->size = 3;
     m_client->Connect(m_addr,false);
 }
